A simple plugin to propagate artifacts across job retries
Leverage the GraphQL API to retrieve previous jobs for a given step,
and list their artifacts.

Filter out artifacts by path, based on a configurable pattern, and
download them into a configurable destination.

This is useful to allow retries to make incremental progress, for
instance this can be used to skip previously successful tests to
save time.