Fix retry for parallel jobs
The BK API does not offer a built-in way to filter jobs by parallel index, which
means the previous query was not guaranteed to pick the previous run of the relevant
job and may instead have gotten the artifacts from a different chunks of parallelized
tests. This was safe, but wasteful, as previously successful tests in this parallel
chunk would therefore not be skipped in the retry.

Update the GraphQL query to fetch 100 jobs instead of 1, and to include the parallelGroupIndex
in the output. Then update the jq query to filter out entries from other parallel indices.

Manually tested on real BK data.